requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pylatex==1.4.2
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.7